+++ core/dive.h
@@ -13,14 +13,13 @@
 #include <stdio.h>
 
 #include "divemode.h"
+#include "divecomputer.h"
 #include "equipment.h"
 #include "picture.h"
 
 #ifdef __cplusplus
 extern "C" {
 #endif
-
-struct event;
 
 extern int last_xml_version;
 
@@ -39,73 +38,13 @@ static inline int interpolate(int a, int b, int part, int whole)
 	return (a+b)/2;
 }
 
-#define MAX_SENSORS 2
-struct sample                         // BASE TYPE BYTES  UNITS    RANGE               DESCRIPTION
-{                                     // --------- -----  -----    -----               -----------
-	duration_t time;                  // int32_t    4  seconds  (0-34 yrs)             elapsed dive time up to this sample
-	duration_t stoptime;              // int32_t    4  seconds  (0-34 yrs)             time duration of next deco stop
-	duration_t ndl;                   // int32_t    4  seconds  (-1 no val, 0-34 yrs)  time duration before no-deco limit
-	duration_t tts;                   // int32_t    4  seconds  (0-34 yrs)             time duration to reach the surface
-	duration_t rbt;                   // int32_t    4  seconds  (0-34 yrs)             remaining bottom time
-	depth_t depth;                    // int32_t    4    mm     (0-2000 km)            dive depth of this sample
-	depth_t stopdepth;                // int32_t    4    mm     (0-2000 km)            depth of next deco stop
-	temperature_t temperature;        // uint32_t   4    mK     (0-4 MK)               ambient temperature
-	pressure_t pressure[MAX_SENSORS]; // int32_t    4    mbar   (0-2 Mbar)             cylinder pressures (main and CCR o2)
-	o2pressure_t setpoint;            // uint16_t   2    mbar   (0-65 bar)             O2 partial pressure (will be setpoint)
-	o2pressure_t o2sensor[3];         // uint16_t   6    mbar   (0-65 bar)             Up to 3 PO2 sensor values (rebreather)
-	bearing_t bearing;                // int16_t    2  degrees  (-1 no val, 0-360 deg) compass bearing
-	uint8_t sensor[MAX_SENSORS];      // uint8_t    1  sensorID (0-255)                ID of cylinder pressure sensor
-	uint16_t cns;                     // uint16_t   1     %     (0-64k %)              cns% accumulated
-	uint8_t heartbeat;                // uint8_t    1  beats/m  (0-255)                heart rate measurement
-	volume_t sac;                     //            4  ml/min                          predefined SAC
-	bool in_deco;                     // bool       1    y/n      y/n                  this sample is part of deco
-	bool manually_entered;            // bool       1    y/n      y/n                  this sample was entered by the user,
-					  //                                               not calculated when planning a dive
-};	                                  // Total size of structure: 57 bytes, excluding padding at end
-
-struct extra_data {
-	const char *key;
-	const char *value;
-	struct extra_data *next;
-};
-
-/*
- * NOTE! The deviceid and diveid are model-specific *hashes* of
- * whatever device identification that model may have. Different
- * dive computers will have different identifying data, it could
- * be a firmware number or a serial ID (in either string or in
- * numeric format), and we do not care.
- *
- * The only thing we care about is that subsurface will hash
- * that information the same way. So then you can check the ID
- * of a dive computer by comparing the hashes for equality.
- *
- * A deviceid or diveid of zero is assumed to be "no ID".
- */
-struct divecomputer {
-	timestamp_t when;
-	duration_t duration, surfacetime, last_manual_time;
-	depth_t maxdepth, meandepth;
-	temperature_t airtemp, watertemp;
-	pressure_t surface_pressure;
-	enum divemode_t divemode;	// dive computer type: OC(default) or CCR
-	uint8_t no_o2sensors;		// rebreathers: number of O2 sensors used
-	int salinity; 			// kg per 10000 l
-	const char *model, *serial, *fw_version;
-	uint32_t deviceid, diveid;
-	int samples, alloc_samples;
-	struct sample *sample;
-	struct event *events;
-	struct extra_data *extra_data;
-	struct divecomputer *next;
-};
-
 struct dive_site;
 struct dive_site_table;
 struct dive_table;
 struct dive_trip;
-struct trip_table;
 struct full_text_cache;
+struct event;
+struct trip_table;
 struct dive {
 	struct dive_trip *divetrip;
 	timestamp_t when;
@@ -171,13 +110,8 @@ struct dive_components {
 	unsigned int weights : 1;
 };
 
-extern enum divemode_t get_current_divemode(const struct divecomputer *dc, int time, const struct event **evp, enum divemode_t *divemode);
-extern struct event *get_next_divemodechange(const struct event **evd, bool update_pointer);
-extern enum divemode_t get_divemode_at_time(const struct divecomputer *dc, int dtime, const struct event **ev_dmc);
-
 extern bool has_gaschange_event(const struct dive *dive, const struct divecomputer *dc, int idx);
 extern int explicit_first_cylinder(const struct dive *dive, const struct divecomputer *dc);
-extern int get_depth_at_time(const struct divecomputer *dc, unsigned int time);
 
 extern fraction_t best_o2(depth_t depth, const struct dive *dive);
 extern fraction_t best_he(depth_t depth, const struct dive *dive, bool o2narcotic, fraction_t fo2);
@@ -190,8 +124,6 @@ extern int rel_mbar_to_depth(int mbar, const struct dive *dive);
 extern int mbar_to_depth(int mbar, const struct dive *dive);
 extern depth_t gas_mod(struct gasmix mix, pressure_t po2_limit, const struct dive *dive, int roundto);
 extern depth_t gas_mnd(struct gasmix mix, depth_t end, const struct dive *dive, int roundto);
-
-#define SURFACE_THRESHOLD 750 /* somewhat arbitrary: only below 75cm is it really diving */
 
 extern bool autogroup;
 
@@ -243,9 +175,6 @@ extern location_t dive_get_gps_location(const struct dive *d);
 
 extern bool time_during_dive_with_offset(const struct dive *dive, timestamp_t when, timestamp_t offset);
 
-/* Check if two dive computer entries are the exact same dive (-1=no/0=maybe/1=yes) */
-extern int match_one_dc(const struct divecomputer *a, const struct divecomputer *b);
-
 extern int save_dives(const char *filename);
 extern int save_dives_logic(const char *filename, bool select_only, bool anonymize);
 extern int save_dive(FILE *f, struct dive *dive, bool anonymize);
@@ -262,19 +191,12 @@ extern bool subsurface_user_is_root(void);
 
 extern struct dive *alloc_dive(void);
 extern void free_dive(struct dive *);
-extern void free_dive_dcs(struct divecomputer *dc);
 extern void record_dive_to_table(struct dive *dive, struct dive_table *table);
 extern void clear_dive(struct dive *dive);
 extern void copy_dive(const struct dive *s, struct dive *d);
 extern void selective_copy_dive(const struct dive *s, struct dive *d, struct dive_components what, bool clear);
 extern struct dive *move_dive(struct dive *s);
 
-extern void alloc_samples(struct divecomputer *dc, int num);
-extern void free_samples(struct divecomputer *dc);
-extern struct sample *prepare_sample(struct divecomputer *dc);
-extern void finish_sample(struct divecomputer *dc);
-extern struct sample *add_sample(const struct sample *sample, int time, struct divecomputer *dc);
-extern void add_sample_pressure(struct sample *sample, int sensor, int mbar);
 extern int legacy_format_o2pressures(const struct dive *dive, const struct divecomputer *dc);
 
 extern bool dive_less_than(const struct dive *a, const struct dive *b);
@@ -284,28 +206,18 @@ extern struct dive *fixup_dive(struct dive *dive);
 extern pressure_t calculate_surface_pressure(const struct dive *dive);
 extern pressure_t un_fixup_surface_pressure(const struct dive *d);
 extern int get_dive_salinity(const struct dive *dive);
-extern void fixup_dc_duration(struct divecomputer *dc);
 extern int dive_getUniqID();
-extern unsigned int dc_airtemp(const struct divecomputer *dc);
-extern unsigned int dc_watertemp(const struct divecomputer *dc);
 extern int split_dive(const struct dive *dive, struct dive **new1, struct dive **new2);
 extern int split_dive_at_time(const struct dive *dive, duration_t time, struct dive **new1, struct dive **new2);
 extern struct dive *merge_dives(const struct dive *a, const struct dive *b, int offset, bool prefer_downloaded, struct dive_trip **trip, struct dive_site **site);
 extern struct dive *try_to_merge(struct dive *a, struct dive *b, bool prefer_downloaded);
-extern void copy_events(const struct divecomputer *s, struct divecomputer *d);
 extern void copy_events_until(const struct dive *sd, struct dive *dd, int time);
 extern void copy_used_cylinders(const struct dive *s, struct dive *d, bool used_only);
-extern void copy_samples(const struct divecomputer *s, struct divecomputer *d);
 extern bool is_cylinder_used(const struct dive *dive, int idx);
 extern bool is_cylinder_prot(const struct dive *dive, int idx);
 extern void add_gas_switch_event(struct dive *dive, struct divecomputer *dc, int time, int idx);
 extern struct event *create_gas_switch_event(struct dive *dive, struct divecomputer *dc, int seconds, int idx);
-extern void add_event_to_dc(struct divecomputer *dc, struct event *ev);
-extern void swap_event(struct divecomputer *dc, struct event *from, struct event *to);
-extern struct event *add_event(struct divecomputer *dc, unsigned int time, int type, int flags, int value, const char *name);
-extern void remove_event_from_dc(struct divecomputer *dc, struct event *event);
 extern void update_event_name(struct dive *d, struct event *event, const char *name);
-extern void add_extra_data(struct divecomputer *dc, const char *key, const char *value);
 extern void per_cylinder_mean_depth(const struct dive *dive, struct divecomputer *dc, int *mean, int *duration);
 extern int get_cylinder_index(const struct dive *dive, const struct event *ev);
 extern struct gasmix get_gasmix_from_event(const struct dive *, const struct event *ev);
@@ -328,7 +240,6 @@ extern const char *existing_filename;
 extern void subsurface_command_line_init(int *, char ***);
 extern void subsurface_command_line_exit(int *, char ***);
 
-extern bool is_dc_planner(const struct divecomputer *dc);
 extern bool has_planned(const struct dive *dive, bool planned);
 
 /* Get gasmixes at increasing timestamps.
@@ -339,13 +250,6 @@ extern struct gasmix get_gasmix(const struct dive *dive, const struct divecomput
 
 /* Get gasmix at a given time */
 extern struct gasmix get_gasmix_at_time(const struct dive *dive, const struct divecomputer *dc, duration_t time);
-
-/* these structs holds the information that
- * describes the cylinders / weight systems.
- * they are global variables initialized in equipment.c
- * used to fill the combobox in the add/edit cylinder
- * dialog
- */
 
 extern void set_informational_units(const char *units);
 extern void set_git_prefs(const char *prefs);
