+++ core/import-divinglog.c
@@ -5,7 +5,9 @@
 #endif
 
 #include "ssrf.h"
+#include "dive.h"
 #include "divesite.h"
+#include "sample.h"
 #include "subsurface-string.h"
 #include "parse.h"
 #include "divelist.h"
