+++ core/import-seac.c
@@ -7,8 +7,8 @@
 #include "qthelper.h"
 #include "ssrf.h"
 #include "dive.h"
+#include "sample.h"
 #include "subsurface-string.h"
-#include "subsurface-time.h"
 #include "parse.h"
 #include "divelist.h"
 #include "device.h"
