+++ core/import-suunto.c
@@ -6,8 +6,9 @@
 
 #include "ssrf.h"
 #include "dive.h"
-#include "subsurface-string.h"
 #include "parse.h"
+#include "sample.h"
+#include "subsurface-string.h"
 #include "divelist.h"
 #include "device.h"
 #include "membuffer.h"
