+++ core/libdivecomputer.c
@@ -14,6 +14,7 @@
 #include <fcntl.h>
 #include "gettext.h"
 #include "divesite.h"
+#include "sample.h"
 #include "subsurface-string.h"
 #include "device.h"
 #include "dive.h"
