+++ core/load-git.c
@@ -18,8 +18,9 @@
 #include "divesite.h"
 #include "event.h"
 #include "errorhelper.h"
-#include "trip.h"
+#include "sample.h"
 #include "subsurface-string.h"
+#include "trip.h"
 #include "device.h"
 #include "membuffer.h"
 #include "git-access.h"
