+++ core/save-git.c
@@ -20,11 +20,13 @@
 #include "divesite.h"
 #include "filterconstraint.h"
 #include "filterpreset.h"
+#include "sample.h"
 #include "subsurface-string.h"
 #include "trip.h"
 #include "device.h"
 #include "errorhelper.h"
 #include "event.h"
+#include "extradata.h"
 #include "membuffer.h"
 #include "git-access.h"
 #include "version.h"
