+++ core/save-xml.c
@@ -15,8 +15,10 @@
 #include "dive.h"
 #include "divesite.h"
 #include "errorhelper.h"
+#include "extradata.h"
 #include "filterconstraint.h"
 #include "filterpreset.h"
+#include "sample.h"
 #include "subsurface-string.h"
 #include "subsurface-time.h"
 #include "trip.h"
