+++ packaging/ios/Subsurface-mobile.pro
@@ -54,6 +54,7 @@ SOURCES += ../../subsurface-mobile-main.cpp \
 	../../core/profile.c \
 	../../core/device.cpp \
 	../../core/dive.c \
+	../../core/divecomputer.c \
 	../../core/divefilter.cpp \
 	../../core/event.c \
 	../../core/filterconstraint.cpp \
@@ -68,6 +69,7 @@ SOURCES += ../../subsurface-mobile-main.cpp \
 	../../core/parse.c \
 	../../core/picture.c \
 	../../core/pictureobj.cpp \
+	../../core/sample.c \
 	../../core/import-suunto.c \
 	../../core/import-shearwater.c \
 	../../core/import-seac.c \
@@ -203,7 +205,9 @@ HEADERS += \
 	../../core/device.h \
 	../../core/devicedetails.h \
 	../../core/dive.h \
+	../../core/divecomputer.h \
 	../../core/event.h \
+	../../core/extradata.h \
 	../../core/git-access.h \
 	../../core/gpslocation.h \
 	../../core/imagedownloader.h \
@@ -240,6 +244,7 @@ HEADERS += \
 	../../core/membuffer.h \
 	../../core/metrics.h \
 	../../core/qt-gui.h \
+	../../core/sample.h \
 	../../core/selection.h \
 	../../core/sha1.h \
 	../../core/strndup.h \
